Add day 16 solution
In hindsight, I'm regretting my decision to use atoms for part 2.
Pretty sure this could be replaced with an idiomatic reduce.
The accumulator for the reduce would be the vector of sets.
Each operation within the reduce would map over that accumulator
and a valid ticket, and do the same basic operation
(find the set of fields we can eliminate from the set
 for that position; use set/difference to do that).
Perhaps an improvement to implement later.